// src/app/model/Board.ts
export interface Board {
    board_idx: number;
    board_num: number;
    board_subject: string;
    board_content: string;
    url_use_yn: string;
    ing_yn: string;
    publish_start_date: string;
    publish_end_date: string;
    push_sended_count: number;
    use_yn: string;
    create_date: string;
    share_board_count: number;
    comment_board_count: 0;
    board_images: null;
    board_image_count: 0;
}